GraphQL IR transforms are incorrect with latest relay-compiler
This brings the transforms applied by typescript in line with flow.
Without this, the codegen is wrong; for example it doesn't generate any
types for fields defined in a client schema.

I think this solution is pretty brittle and likely to break in the
future so it's probably best to work out a better way of accessing the
transforms needed by the typescript compiler in the long term